Move error functions into error module
